Revert "Sp5 profile alex"
