+++ fytProject/components/containers/WorkoutContainer.js
@@ -1,20 +1,13 @@
 import React from 'react';
 import {bindActionCreators} from 'redux';
-import {View} from 'react-native';
 import { connect } from 'react-redux';
 import WorkoutList from '../WorkoutList.js';
-import {calculateWorkoutProgress} from '../../config/utilities.js';
-
 import {editSet, changeSetCompletionStatus} from '../../actions/WorkoutActions';
 
 class WorkoutContainer extends React.Component {
 
-
   render(){
-    let workoutProgress = calculateWorkoutProgress(this.props.workout);
-    return(
-      <WorkoutList workoutProgress={workoutProgress} workout={this.props.workout}/>
-  );
+    return(<WorkoutList workout={this.props.workout}/>);
   }
 
 }
