+++ fytProject/package.json
@@ -9,7 +9,6 @@
     "test": "jest"
   },
   "dependencies": {
-    "@expo/vector-icons": "^8.0.0",
     "axios": "^0.18.0",
     "expo": "^31.0.2",
     "native-base": "2.8.1",
@@ -18,15 +17,18 @@
     "react-devtools": "^3.4.3",
     "react-native": "^0.57.8",
     "react-native-debugger": "^1.1.0",
-    "react-native-progress-bar-animated": "^1.0.6",
-    "react-navigation": "1.5.1",
     "react-redux": "^6.0.0",
     "redux": "^4.0.1",
     "redux-axios-middleware": "^4.0.0",
     "redux-thunk": "^2.3.0",
-    "tcomb-form-native": "^0.6.20"
+    "tcomb-form-native": "^0.6.20",
+    "@expo/vector-icons": "^8.0.0",
+    "react-navigation": "1.5.1"
   },
   "devDependencies": {
+    "babel-jest": "^23.6.0",
+    "babel-preset-expo": "^5.0.0",
+    "jest": "^23.6.0",
     "react-native-debugger-open": "^0.3.17",
     "react-test-renderer": "^16.6.3"
   },
